+++ adm/app/adms/Views/sitsUsers/listSitsUsers.php
@@ -16,7 +16,7 @@ foreach($this->data['listSitsUsers'] as $listSits){
     echo "Situação e Cor: <span style='color:$color;'> $sitsname</span><br><br>";
     echo "<a class='btn btn-sm btn-outline-primary ms-4' href='".URLADM."view-sits-users/index/$id'> Visualizar </a>";
     echo "<a class='btn btn-sm btn-outline-primary ms-4' href='".URLADM."edit-sits-users/index/$id'> Editar </a>";
-    echo "<a class='btn btn-sm btn-outline-primary ms-4' href='".URLADM."delete-sits-users/index/$id'> Apagar </a><br><hr>";
+    echo "<a class='btn btn-sm btn-outline-primary ms-4' href='".URLADM."delete-sits-users/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir o registro?\")'> Apagar </a><br><hr>";
 }
 //imprime os links de paginação
 echo $this->data['pagination'];
